+++ server/src/main/java/org/opensearch/node/Node.java
@@ -436,6 +436,9 @@ public class Node implements Closeable {
 
             final Settings settings = pluginsService.updatedSettings();
 
+            // Ensure to initialise Feature Flags via the settings from opensearch.yml
+            FeatureFlags.initialiseFeatureFlags(settings);
+
             final Set<DiscoveryNodeRole> additionalRoles = pluginsService.filterPlugins(Plugin.class)
                 .stream()
                 .map(Plugin::getRoles)